requirements.txt
redis>=4.5.0
python-dotenv>=1.0.0 
aiohttp>=3.8.4
psycopg2-binary>=2.9.6
psycopg2-pool>=1.2
boto3>=1.34.0